requirements.txt
streamlit
pandas
scikit-learn
nltk
pdfplumber
python-docx